requirements.txt
aiosmtpd==1.4.4.post2
asgiref==3.6.0
atpublic==4.0
attrs==23.1.0
distlib==0.3.6
dj-database-url==2.0.0
Django==4.1.7
django-ckeditor==6.5.1
django-heroku==0.3.1
django-js-asset==2.0.0
filelock==3.10.0
gunicorn==21.2.0
packaging==23.1
Pillow==9.5.0
platformdirs==3.1.1
psycopg2==2.9.6
python-decouple==3.8
pytz==2023.3
sqlparse==0.4.3
typing_extensions==4.7.1
tzdata==2022.7
virtualenv==20.21.0
virtualenvwrapper-win==1.2.7
whitenoise==6.5.0
